+++ .gitignore
@@ -1,0 +1,10 @@
+
+.vs/DiscordBot/v16/.suo
+
+.vs/ProjectSettings.json
+
+.vs/slnx.sqlite
+
+.vs/DiscordBot/config/applicationhost.config
+
+.vs/VSWorkspaceState.json
