+++ commands/tn.js
@@ -1,15 +1,11 @@
 const { MessageEmbed, splitMessage } = require("discord.js");
 const config = require('../config.js');
 
-module.exports = {
-    name: "startfl",
-    aliases: ['fl'],
-    description: "Check FL warehouse folders.",
-    execute(message) {
+exports.run = (client, message, args) => {
 
         // Imports / Requires
-        var dirwatch = require("./DirectoryWatcher.js");
-        var flChannel = globalClient.channels.get(config.locations.FL);
+        var dirwatch = require("../modules/DirectoryWatcher.js");
+        var tnChannel = globalClient.channels.get(config.locations.TN);
 
         // Create a monitor object that will watch a directory
         // and all it's sub-directories (recursive) in this case
@@ -20,15 +16,15 @@ module.exports = {
         // you can monitor only a single folder and none of its child
         // directories by simply changing the recursive parameter to
         // to false
-        var flMonitor = new dirwatch.DirectoryWatcher("Z:\\10A-Processed", true);
+        var tnMonitor = new dirwatch.DirectoryWatcher("Z:\\13A-Processed", true);
 
         // start the monitor and have it check for updates
         // every half second.
-        flMonitor.start(60000);
+        tnMonitor.start(60000);
         
         // Log to the console when a file is removed
-        flMonitor.on("fileRemoved", function (filePath) {
-            flChannel.send({
+        tnMonitor.on("fileRemoved", function (filePath) {
+            tnChannel.send({
                 embed: {
                     color: 0x2ecc71,
                     title: "PickTicket Deleted",
@@ -47,8 +43,8 @@ module.exports = {
         });
 
         // Log to the console when a folder is removed
-        flMonitor.on("folderRemoved", function (folderPath) {
-            flChannel.send({
+        tnMonitor.on("folderRemoved", function (folderPath) {
+            tnChannel.send({
                 embed: {
                     color: 0x2ecc71,
                     title: "Folder Deleted",
@@ -67,8 +63,8 @@ module.exports = {
         });
 
         // log to the console when a folder is added
-        flMonitor.on("folderAdded", function (folderPath) {
-            flChannel.send({
+        tnMonitor.on("folderAdded", function (folderPath) {
+            tnChannel.send({
                 embed: {
                     color: 0x2ecc71,
                     title: "Folder Added",
@@ -87,9 +83,9 @@ module.exports = {
         });
 
         // Log to the console when a file is changed.
-        flMonitor.on("fileChanged", function (fileDetail, changes) {
+        tnMonitor.on("fileChanged", function (fileDetail, changes) {
             for (var key in changes) {
-                flChannel.send({
+                tnChannel.send({
                     embed: {
                         color: 0x2ecc71,
                         title: "File Changed",
@@ -118,8 +114,8 @@ module.exports = {
         });
 
         // log to the console when a file is added.
-        flMonitor.on("fileAdded", function (fileDetail) {            
-            flChannel.send({
+        tnMonitor.on("fileAdded", function (fileDetail) {            
+            tnChannel.send({
                 embed: {
                     color: 0x2ecc71,
                     title: "New PickTicket",
@@ -137,13 +133,13 @@ module.exports = {
             console.log("File Added: " + fileDetail.fullPath);
         });
 
-        flChannel.send({
+        tnChannel.send({
             embed: {
                 color: 0x2ecc71,
-                title: "Monitoring Florida PickTicket Folder!",
+                title: "Monitoring Tennessee PickTicket Folder!",
                 fields: [{
                     name: "Monitoring mapped drive below!:",
-                    value: flMonitor.root
+                    value: tnMonitor.root
                 }
                 ],
                 timestamp: new Date(),
@@ -153,7 +149,7 @@ module.exports = {
             }
         });
         // Let us know that directory monitoring is happening and where.
-        console.log("Directory Monitoring of " + flMonitor.root + " has started");
+        console.log("Directory Monitoring of " + tnMonitor.root + " has started");
 
     }
 };
