+++ events/error.js
@@ -1,0 +1,3 @@
+module.exports = async (client, error) => {
+    client.logger.log(`An error event was sent by Discord.js: \n${JSON.stringify(error)}`, "error");
+};
