+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "discord.js": "^11.4.2",
     "pm2": "^2.10.4",
-    "request-promise": "^4.2.4"
+    "request-promise": "^4.2.4",
+    "enmap": "5.3.1"
   }
 }
